- Use string instead of char type

+++ calculator/calculator-logic/src/main/java/calculator/Calculator.java
@@ -5,16 +5,16 @@ public class Calculator {
     public Calculator() {
     }
     
-    public double calculate(String operator, int a, int b) {
+    public double calculate(String operator, double a, double b) {
         double result = 0;
         
-        if (operator.equals('+')) {
+        if (operator.equals("+")) {
             result = a + b;
-        } else if (operator.equals('-')) {
+        } else if (operator.equals("-")) {
             result = a - b;
-        } else if (operator.equals('*')) {
+        } else if (operator.equals("*")) {
             result = a * b;
-        } else if (operator.equals('/')) {
+        } else if (operator.equals("/")) {
             result = a / b;
         }
